requirements.txt
playwright
httpx
python-dateutil
tenacity
ultralytics
opencv-python
numpy
flask

